[FeatureD] Move the Feature related config from Hostcfgd into a new daemon (#71)
### Description

Move feature related config from hostcfgd into a new daemon

```
tests/featured/featured_test.py::TestFeatureHandler::test_feature_config_parsing PASSED [ 46%]
tests/featured/featured_test.py::TestFeatureHandler::test_feature_config_parsing_defaults PASSED [ 47%]
tests/featured/featured_test.py::TestFeatureHandler::test_feature_resync PASSED [ 48%]
tests/featured/featured_test.py::TestFeatureHandler::test_handler_00_DualTorCase <- ../../../usr/local/lib/python3.9/dist-packages/pyfakefs/fake_filesystem_unittest.py PASSED [ 49%]
tests/featured/featured_test.py::TestFeatureHandler::test_handler_01_SingleToRCase <- ../../../usr/local/lib/python3.9/dist-packages/pyfakefs/fake_filesystem_unittest.py PASSED [ 50%]
tests/featured/featured_test.py::TestFeatureHandler::test_handler_02_T1Case <- ../../../usr/local/lib/python3.9/dist-packages/pyfakefs/fake_filesystem_unittest.py PASSED [ 50%]
tests/featured/featured_test.py::TestFeatureHandler::test_handler_03_SingleToRCase_DHCP_Relay_Enabled <- ../../../usr/local/lib/python3.9/dist-packages/pyfakefs/fake_filesystem_unittest.py PASSED [ 51%]
tests/featured/featured_test.py::TestFeatureHandler::test_handler_04_DualTorCaseWithNoSystemCalls <- ../../../usr/local/lib/python3.9/dist-packages/pyfakefs/fake_filesystem_unittest.py PASSED [ 52%]
tests/featured/featured_test.py::TestFeatureHandler::test_handler_05_Chassis_Supervisor_PACKET <- ../../../usr/local/lib/python3.9/dist-packages/pyfakefs/fake_filesystem_unittest.py PASSED [ 53%]
tests/featured/featured_test.py::TestFeatureHandler::test_handler_06_Chassis_Supervisor_VOQ <- ../../../usr/local/lib/python3.9/dist-packages/pyfakefs/fake_filesystem_unittest.py PASSED [ 54%]
tests/featured/featured_test.py::TestFeatureHandler::test_handler_07_Chassis_LineCard_VOQ <- ../../../usr/local/lib/python3.9/dist-packages/pyfakefs/fake_filesystem_unittest.py PASSED [ 55%]
tests/featured/featured_test.py::TestFeatureHandler::test_handler_08_Chassis_LineCard_Packet <- ../../../usr/local/lib/python3.9/dist-packages/pyfakefs/fake_filesystem_unittest.py PASSED [ 56%]
tests/featured/featured_test.py::TestFeatureHandler::test_handler_09_Chassis_Supervisor_PACKET_multinpu <- ../../../usr/local/lib/python3.9/dist-packages/pyfakefs/fake_filesystem_unittest.py PASSED [ 56%]
tests/featured/featured_test.py::TestFeatureHandler::test_handler_10_Chassis_LineCard_VOQ_multinpu <- ../../../usr/local/lib/python3.9/dist-packages/pyfakefs/fake_filesystem_unittest.py PASSED [ 57%]
tests/featured/featured_test.py::TestFeatureHandler::test_sync_state_field_00_DualTorCase <- ../../../usr/local/lib/python3.9/dist-packages/pyfakefs/fake_filesystem_unittest.py PASSED [ 58%]
tests/featured/featured_test.py::TestFeatureHandler::test_sync_state_field_01_SingleToRCase <- ../../../usr/local/lib/python3.9/dist-packages/pyfakefs/fake_filesystem_unittest.py PASSED [ 59%]
tests/featured/featured_test.py::TestFeatureHandler::test_sync_state_field_02_T1Case <- ../../../usr/local/lib/python3.9/dist-packages/pyfakefs/fake_filesystem_unittest.py PASSED [ 60%]
tests/featured/featured_test.py::TestFeatureHandler::test_sync_state_field_03_SingleToRCase_DHCP_Relay_Enabled <- ../../../usr/local/lib/python3.9/dist-packages/pyfakefs/fake_filesystem_unittest.py PASSED [ 61%]
tests/featured/featured_test.py::TestFeatureHandler::test_sync_state_field_04_DualTorCaseWithNoSystemCalls <- ../../../usr/local/lib/python3.9/dist-packages/pyfakefs/fake_filesystem_unittest.py PASSED [ 62%]
tests/featured/featured_test.py::TestFeatureHandler::test_sync_state_field_05_Chassis_Supervisor_PACKET <- ../../../usr/local/lib/python3.9/dist-packages/pyfakefs/fake_filesystem_unittest.py PASSED [ 62%]
tests/featured/featured_test.py::TestFeatureHandler::test_sync_state_field_06_Chassis_Supervisor_VOQ <- ../../../usr/local/lib/python3.9/dist-packages/pyfakefs/fake_filesystem_unittest.py PASSED [ 63%]
tests/featured/featured_test.py::TestFeatureHandler::test_sync_state_field_07_Chassis_LineCard_VOQ <- ../../../usr/local/lib/python3.9/dist-packages/pyfakefs/fake_filesystem_unittest.py PASSED [ 64%]
tests/featured/featured_test.py::TestFeatureHandler::test_sync_state_field_08_Chassis_LineCard_Packet <- ../../../usr/local/lib/python3.9/dist-packages/pyfakefs/fake_filesystem_unittest.py PASSED [ 65%]
tests/featured/featured_test.py::TestFeatureHandler::test_sync_state_field_09_Chassis_Supervisor_PACKET_multinpu <- ../../../usr/local/lib/python3.9/dist-packages/pyfakefs/fake_filesystem_unittest.py PASSED [ 66%]
tests/featured/featured_test.py::TestFeatureHandler::test_sync_state_field_10_Chassis_LineCard_VOQ_multinpu <- ../../../usr/local/lib/python3.9/dist-packages/pyfakefs/fake_filesystem_unittest.py PASSED [ 67%]
tests/featured/featured_test.py::TestFeatureDaemon::test_advanced_reboot PASSED [ 68%]
tests/featured/featured_test.py::TestFeatureDaemon::test_delayed_service PASSED [ 68%]
tests/featured/featured_test.py::TestFeatureDaemon::test_feature_events PASSED [ 69%]
tests/featured/featured_test.py::TestFeatureDaemon::test_portinit_timeout PASSED [ 70%]
```
